// frontend/src/types/auth.ts
export type ActionMap<M extends { [index: string]: any }> = {
  [Key in keyof M]: M[Key] extends undefined
  ? {
    type: Key;
  }
  : {
    type: Key;
    payload: M[Key];
  };
};

export type AuthUser = null | Record<string, any>;

export type AuthState = {
  accessToken: string | null;
  refreshToken: string | null;
  user: AuthUser;
};

export type JWTContextType = {
  isAuthenticated: boolean;
  isInitialized: boolean;
  user: AuthUser;
  method: "jwt";
  signIn: (email: string, password: string) => Promise<void>;
  signOut: () => Promise<void>;
  // signUp: (
  //   email: string,
  //   password: string,
  //   firstName: string,
  //   lastName: string
  // ) => Promise<void>;
  resetPassword: (email: string) => void;
};